requirements.txt
Flask==1.0.2
mysql-connector-python==8.0.12
Jinja2==2.10
pytz==2018.5